Move `testIds` to `renderer/`
