+++ src/renderer/App/AppBar/SearchBox/SearchBoxHooks.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useHistory } from "react-router-dom";
-import { RpcService } from "_r/RpcClient/RpcService";
+import { RpcService } from "_r/RpcService/RpcService";
 import { BITCOIN_CORE_RPC_ERROR } from "_c/constants";
 
 const hashRegex = /[0-9a-fA-F]{64}/;
