+++ src/renderer/App/AppBar/SearchBox/tests/SearchBox.test.ts
@@ -5,7 +5,7 @@ import { initializeElectronCode } from "_tu/initializeElectronCode";
 import { renderAppWithStore } from "_tu/renderAppWithStore";
 import { waitWithTime } from "_tu/smallUtils";
 import { startMockRpcServer } from "_tu/startMockRpcServer";
-import { testIds } from "_tu/testIds";
+import { testIds } from "_r/testIds";
 
 jest.mock("_f/featureFlags", () => ({
   __esModule: true,
